new changes

+++ database/migrations/2024_08_28_111617_add_new_fields_to_characters.php
@@ -16,7 +16,7 @@ class AddNewFieldsToCharacters extends Migration
         Schema::table('characters', function (Blueprint $table) {
             $table->integer('initial_age')->nullable();
             $table->integer('age')->nullable();
-            $table->integer('height')->nullable();
+            $table->decimal('height', 4, 1)->nullable();
             $table->jsonb('nickname')->nullable();
         });
     }
